+++ packages/ts-cleanco/cleanco.ts
@@ -1,4 +1,0 @@
-export type CleanCo = {
-  name: string;
-  sound: string;
-};