+++ CHANGELOG.md
@@ -1,3 +1,125 @@
+## 1.1.15 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.14 (2024-11-17)
+
+### 🚀 Features
+
+- Fallback company logo fetch ([a708f03](https://github.com/qonstrue/opensource/commit/a708f03))
+
+### ❤️  Thank You
+
+- sustentabilitas
+
+## 1.1.13 (2024-11-17)
+
+### 🚀 Features
+
+- add fetchAsBase64String to download logos as base64 using fetch ([448dac7](https://github.com/qonstrue/opensource/commit/448dac7))
+
+### ❤️  Thank You
+
+- sustentabilitas
+
+## 1.1.12 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.11 (2024-11-17)
+
+### 🚀 Features
+
+- Added package to extract company logos or favicons ([6e38605](https://github.com/qonstrue/opensource/commit/6e38605))
+
+### ❤️  Thank You
+
+- sustentabilitas
+
+## 1.1.10 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.9 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.8 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.7 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.6 (2024-11-17)
+
+### 🚀 Features
+
+- Handle punctuation in replacements at the start of the name ([b171c7d](https://github.com/qonstrue/opensource/commit/b171c7d))
+
+### ❤️  Thank You
+
+- sustentabilitas
+
+## 1.1.5 (2024-11-17)
+
+### 🚀 Features
+
+- additional test cases from the python library ([3851bcf](https://github.com/qonstrue/opensource/commit/3851bcf))
+
+### ❤️  Thank You
+
+- sustentabilitas
+
+## 1.1.4 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.3 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.2 (2024-11-17)
+
+This was a version bump only, there were no code changes.
+
+## 1.1.1 (2024-11-16)
+
+### 🚀 Features
+
+- **python:** support what the python code does ([488242d](https://github.com/qonstrue/opensource/commit/488242d))
+
+### ❤️  Thank You
+
+- sustentabilitas
+
+## 1.1.0 (2024-11-16)
+
+### 🚀 Features
+
+- **python:** support what the python code does ([488242d](https://github.com/qonstrue/opensource/commit/488242d))
+
+### ❤️  Thank You
+
+- sustentabilitas
+
+## 1.0.4 (2024-11-16)
+
+This was a version bump only, there were no code changes.
+
+## 1.0.3 (2024-11-16)
+
+This was a version bump only, there were no code changes.
+
+## 1.0.2 (2024-11-16)
+
+This was a version bump only, there were no code changes.
+
+## 1.0.1 (2024-11-16)
+
+This was a version bump only, there were no code changes.
+
 ## 1.2.1 (2024-11-15)
 
 This was a version bump only, there were no code changes.